Cleanup root namespace
