+++ pytensor/tensor/rewriting/ofg.py
@@ -1,9 +1,8 @@
 from typing import cast
 
-from pytensor import Variable, clone_replace
 from pytensor.compile import optdb
 from pytensor.compile.builders import OpFromGraph
-from pytensor.graph import Apply, node_rewriter
+from pytensor.graph import Apply, Variable, clone_replace, node_rewriter
 from pytensor.graph.rewriting.basic import copy_stack_trace, dfs_rewriter
 from pytensor.tensor.basic import AllocDiag
 from pytensor.tensor.rewriting.basic import register_specialize
