+++ pytensor/tensor/rewriting/subtensor_lift.py
@@ -4,9 +4,14 @@ from typing import cast
 import numpy as np
 from numpy.lib.array_utils import normalize_axis_index, normalize_axis_tuple
 
-from pytensor import Variable
 from pytensor.compile import optdb
-from pytensor.graph import Constant, FunctionGraph, node_rewriter, vectorize_graph
+from pytensor.graph import (
+    Constant,
+    FunctionGraph,
+    Variable,
+    node_rewriter,
+    vectorize_graph,
+)
 from pytensor.graph.rewriting.basic import NodeRewriter, copy_stack_trace
 from pytensor.tensor.basic import (
     Alloc,
