+++ pytensor/tensor/type_other.py
@@ -5,7 +5,7 @@
 import numpy as np
 
 import pytensor
-from pytensor import _as_symbolic
+from pytensor.basic import _as_symbolic
 from pytensor.gradient import disconnected_type
 from pytensor.graph.basic import Apply, Constant, Variable
 from pytensor.graph.op import Op
