+++ scripts/mypy-failing.txt
@@ -23,6 +23,5 @@ pytensor/tensor/random/op.py
 pytensor/tensor/random/utils.py
 pytensor/tensor/rewriting/basic.py
 pytensor/tensor/type.py
-pytensor/tensor/type_other.py
 pytensor/tensor/variable.py
 pytensor/_version.py
