+++ tests/test_basic.py
@@ -1,0 +1,32 @@
+import pytensor
+
+
+def test_root_module_not_polluted():
+    import types
+
+    # Filter out submodules since other tests may have imported them
+    module_items = sorted(
+        i
+        for i in dir(pytensor)
+        if not i.startswith("__")
+        and not isinstance(getattr(pytensor, i), types.ModuleType)
+    )
+    assert module_items == [
+        "In",
+        "Lop",
+        "Mode",
+        "OpFromGraph",
+        "Out",
+        "Rop",
+        "config",
+        "dprint",
+        "function",
+        "get_mode",
+        "grad",
+        "ifelse",
+        "map",
+        "scan",
+        "shared",
+        "wrap_jax",
+        "wrap_py",
+    ]
